+++ src/components/SegmentedControl/index.ts
@@ -1,2 +1,1 @@
 export * from "./SegmentedControl";
-export * from "./variants";